Revert "Move component files out of glasslibrary package"
This reverts commit 43526df3
