+++ src/org/labkey/test/components/glassLibrary/ParentEntityEditPanel.java
@@ -5,8 +5,8 @@ import org.labkey.test.Locator;
 import org.labkey.test.WebDriverWrapper;
 import org.labkey.test.components.Component;
 import org.labkey.test.components.WebDriverComponent;
-import org.labkey.test.components.react.select.FilteringReactSelect;
-import org.labkey.test.components.react.select.ReactSelect;
+import org.labkey.test.components.glassLibrary.components.FilteringReactSelect;
+import org.labkey.test.components.glassLibrary.components.ReactSelect;
 import org.openqa.selenium.StaleElementReferenceException;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
