+++ src/org/labkey/test/components/glassLibrary/components/FilteringReactSelect.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2018-2019 LabKey Corporation. All rights reserved. No portion of this work may be reproduced in
  * any form or by any electronic or mechanical means without written permission from LabKey Corporation.
  */
-package org.labkey.test.components.react.select;
+package org.labkey.test.components.glassLibrary.components;
 
 import org.labkey.test.Locator;
 import org.labkey.test.WebDriverWrapper;
