+++ src/org/labkey/test/components/glassLibrary/files/FileSelectTree.java
@@ -1,4 +1,4 @@
-package org.labkey.test.components.react.files;
+package org.labkey.test.components.glassLibrary.files;
 
 import org.labkey.test.Locator;
 import org.labkey.test.WebDriverWrapper;
