+++ src/org/labkey/test/components/glassLibrary/grids/GridTabBar.java
@@ -1,4 +1,4 @@
-package org.labkey.test.components.react.grids;
+package org.labkey.test.components.glassLibrary.grids;
 
 import org.labkey.test.Locator;
 import org.labkey.test.WebDriverWrapper;
