+++ src/org/labkey/test/components/list/AdvancedListSettingsDialog.java
@@ -3,9 +3,9 @@ package org.labkey.test.components.list;
 import org.labkey.test.Locator;
 import org.labkey.test.WebDriverWrapper;
 import org.labkey.test.components.bootstrap.ModalDialog;
+import org.labkey.test.components.glassLibrary.components.ReactSelect;
 import org.labkey.test.components.html.Input;
 import org.labkey.test.components.html.RadioButton;
-import org.labkey.test.components.react.select.ReactSelect;
 import org.labkey.test.pages.core.login.SvgCheckbox;
 import org.labkey.test.pages.list.EditListDefinitionPage;
 import org.openqa.selenium.WebElement;
