+++ src/org/labkey/test/components/ui/EntityBulkUpdateDialog.java
@@ -2,10 +2,10 @@ package org.labkey.test.components.ui;
 
 import org.labkey.test.Locator;
 import org.labkey.test.components.bootstrap.ModalDialog;
+import org.labkey.test.components.glassLibrary.components.FilteringReactSelect;
 import org.labkey.test.components.html.Checkbox;
 import org.labkey.test.components.html.Input;
 import org.labkey.test.components.html.ToggleButton;
-import org.labkey.test.components.react.select.FilteringReactSelect;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
