+++ src/org/labkey/test/components/ui/EntityInsertPanel.java
@@ -3,8 +3,8 @@ package org.labkey.test.components.ui;
 import org.labkey.test.Locator;
 import org.labkey.test.components.Component;
 import org.labkey.test.components.WebDriverComponent;
-import org.labkey.test.components.react.grids.EditableGrid;
-import org.labkey.test.components.react.select.ReactSelect;
+import org.labkey.test.components.glassLibrary.components.ReactSelect;
+import org.labkey.test.components.glassLibrary.grids.EditableGrid;
 import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
